Extract linker arg parser to linker crate
